Fix : DIG-76 Value 값 주입 못하는 버그 수정 및 잘못된 상수값 수정
- LoginResponse가 빈으로 등록되지 않아서 Value 어노테이션이 동작 안하고 있었다.
- NOT_ALREADY_JOINED 잘못된 값 수정
